vlm_transformers_requirements.txt
transformers>=4.40.0
torch>=2.0.0
torchvision>=0.15.0
pillow>=10.0.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
python-dotenv>=1.0.0
accelerate>=0.24.0
